pyproject.toml
[tool.poetry]
name = "Device_Anomaly_Detection"
version = "0.1.0"
description = ""
authors = ["WisamAbbasi <39978331+WisamAbbasi@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
pytest = "^7.2.1"
tensorflow = "^2.12.1, <3.0.0"
tensorflow-privacy = "0.8.10"
flask = "2.3.2"
numpy = "1.24.3"
websocket-client = "1.6.1"
rel = "0.4.9"

[tool.poetry.dev-dependencies]
pytest = "^7.2.1"
pre-commit = "^3.0.4"
black = "^23.1.0"
isort = "^5.12.0"
ruff = "^0.0.63"
coverage = {extras = ["toml"], version = "^7.1.0"}

[build-system]
requires = ["poetry-core>=1.0.0", "pip>=22.1.2"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79

[tool.isort]
profile = "black"
line_length = 79

[tool.ruff]
line-length = 79
select = [
    "F401",
    "F403",
]

[tool.coverage.run]
omit = [".*", "*/site-packages/*"]

[tool.coverage.report]
fail_under = 60